bug 1 fixed- dropping labels
in this commit the bug 1 fix issue that entailed the labels being able to drop in the same drop zone were fixed- this was fixed by changing the if and this statements in the drop event. the this.querySelector(".label" checks if a label has already been dropped and if yes the clicked label returns back to the label menu.